requirements.txt
torch 
torchvision
deepforest
fastapi 
uvicorn 
opencv-python 
Pillow 
numpy
geopandas
python-multipart